requirements.txt
PyMuPDF
Pillow
zxing-cpp